feat(be): create lambda layer for perm methods, improve protections for endpoints, organize permission endpoints into new lambda
