+++ aws/lambdas/strategies/main.py
@@ -9,6 +9,12 @@ from boto3.dynamodb.types import TypeSerializer
 from boto3.dynamodb.conditions import Key
 from botocore.exceptions import ClientError
 
+from hqg_permissions import (
+    get_roles_from_event,
+    has_read_permission,
+    role_principals,
+)
+
 dynamo = boto3.resource("dynamodb")
 dynamo_client = boto3.client("dynamodb")
 s3 = boto3.client("s3")
@@ -38,9 +44,6 @@ def handler(event: Dict[str, Any], _context: Any) -> Dict[str, Any]:
     - POST /strategies
     - GET /strategies/{id}
     - PATCH /strategies/{id}
-    - GET /strategies/{id}/permissions
-    - PATCH /strategies/{id}/permissions
-    - DELETE /strategies/{id}/permissions
     """
     route_key = (event.get("requestContext") or {}).get("routeKey")
 
@@ -60,20 +63,6 @@ def handler(event: Dict[str, Any], _context: Any) -> Dict[str, Any]:
         body = json.loads(event.get("body") or "{}")
         return update_strategy(strategy_id, body)
 
-    if route_key == "GET /strategies/{id}/permissions":
-        strategy_id = (event.get("pathParameters") or {}).get("id")
-        return get_permissions(strategy_id, event)
-
-    if route_key == "PATCH /strategies/{id}/permissions":
-        strategy_id = (event.get("pathParameters") or {}).get("id")
-        body = json.loads(event.get("body") or "{}")
-        return update_permissions(strategy_id, body, event, mode="patch")
-
-    if route_key == "DELETE /strategies/{id}/permissions":
-        strategy_id = (event.get("pathParameters") or {}).get("id")
-        body = json.loads(event.get("body") or "{}")
-        return update_permissions(strategy_id, body, event, mode="delete")
-
     return {"statusCode": 404, "body": "Not found"}
 
 
@@ -82,7 +71,7 @@ def get_strategies(event: Dict[str, Any]) -> Dict[str, Any]:
     if not netid:
         return _json(401, {"message": "unauthorized"})
 
-    roles = _get_roles_from_event(event)
+    roles = get_roles_from_event(event)
     if "ADMIN" in roles:
         resp = STRATEGIES_TABLE.scan()
         items = _clean_decimals(resp.get("Items", []))
@@ -167,12 +156,17 @@ def create_strategy(body: Dict[str, Any], event: Dict[str, Any]) -> Dict[str, An
     netid = _get_netid_from_event(event)
     if not netid:
         return _json(401, {"message": "unauthorized"})
+    roles = get_roles_from_event(event)
     owner = netid
     owner_display = _get_display_name_from_event(event) or netid
 
     source = _get_strategy_by_id(source_id)
     if not source:
         return _json(404, {"message": f"Source strategy {source_id} not found"})
+    if "ADMIN" not in roles and not has_read_permission(
+        source_id, netid, roles, READ_PERMISSIONS_DDB, WRITE_PERMISSIONS_DDB
+    ):
+        return _json(403, {"message": "forbidden"})
 
     new_strategy_id = str(_next_strategy_id())
     new_version = 1
@@ -222,14 +216,6 @@ def create_strategy(body: Dict[str, Any], event: Dict[str, Any]) -> Dict[str, An
                 },
                 {
                     "Put": {
-                        "TableName": READ_PERMISSIONS_TABLE,
-                        "Item": _ddb_item(
-                            {"strategy_id": new_strategy_id, "principal": "ROLE#PUBLIC"}
-                        ),
-                    }
-                },
-                {
-                    "Put": {
                         "TableName": WRITE_PERMISSIONS_TABLE,
                         "Item": _ddb_item(
                             {"strategy_id": new_strategy_id, "principal": owner_principal}
@@ -252,8 +238,10 @@ def get_strategy(strategy_id: Optional[str], event: Dict[str, Any]) -> Dict[str,
     if not netid:
         return _json(401, {"message": "unauthorized"})
 
-    roles = _get_roles_from_event(event)
-    if "ADMIN" not in roles and not _has_read_permission(strategy_id, netid, roles):
+    roles = get_roles_from_event(event)
+    if "ADMIN" not in roles and not has_read_permission(
+        strategy_id, netid, roles, READ_PERMISSIONS_DDB, WRITE_PERMISSIONS_DDB
+    ):
         return _json(403, {"message": "forbidden"})
 
     item = _get_strategy_by_id(strategy_id)
@@ -272,167 +260,6 @@ def update_strategy(strategy_id: Optional[str], body: Dict[str, Any]) -> Dict[st
     item = {"id": strategy_id, **body, "updated_at": now}
     STRATEGIES_TABLE.put_item(Item=item)
     return _json(200, item)
-
-
-def get_permissions(strategy_id: Optional[str], event: Dict[str, Any]) -> Dict[str, Any]:
-    if not strategy_id:
-        return _json(400, {"message": "strategy id is required"})
-
-    netid = _get_netid_from_event(event)
-    if not netid:
-        return _json(401, {"message": "unauthorized"})
-
-    strategy = _get_strategy_by_id(strategy_id)
-    if not strategy:
-        return _json(404, {"message": "Strategy not found"})
-
-    roles = _get_roles_from_event(event)
-    owner = strategy.get("owner")
-    if "ADMIN" not in roles and (not owner or owner != netid):
-        return _json(403, {"message": "forbidden"})
-
-    return _json(
-        200,
-        {
-            "read": _permission_snapshot(READ_PERMISSIONS_DDB, strategy_id),
-            "write": _permission_snapshot(WRITE_PERMISSIONS_DDB, strategy_id),
-        },
-    )
-
-
-def update_permissions(
-    strategy_id: Optional[str],
-    body: Dict[str, Any],
-    event: Dict[str, Any],
-    mode: str,
-) -> Dict[str, Any]:
-    if not strategy_id:
-        return _json(400, {"message": "strategy id is required"})
-
-    netid = _get_netid_from_event(event)
-    if not netid:
-        return _json(401, {"message": "unauthorized"})
-
-    strategy = _get_strategy_by_id(strategy_id)
-    if not strategy:
-        return _json(404, {"message": "Strategy not found"})
-
-    roles = _get_roles_from_event(event)
-    owner = strategy.get("owner")
-    if "ADMIN" not in roles and (not owner or owner != netid):
-        return _json(403, {"message": "forbidden"})
-
-    if not isinstance(body, dict):
-        return _json(400, {"message": "body must be an object"})
-
-    for scope, table in (("read", READ_PERMISSIONS_DDB), ("write", WRITE_PERMISSIONS_DDB)):
-        if scope not in body:
-            continue
-        scope_body = body.get(scope)
-        if not isinstance(scope_body, dict):
-            return _json(400, {"message": f"{scope} must be an object"})
-        error = _apply_permission_changes(table, strategy_id, scope_body, mode)
-        if error:
-            return error
-        if scope == "write" and mode == "patch":
-            implied_read = _read_implied_changes(scope_body)
-            if implied_read:
-                error = _apply_permission_changes(
-                    READ_PERMISSIONS_DDB, strategy_id, implied_read, "patch"
-                )
-                if error:
-                    return error
-
-    return _json(200, {"ok": True})
-
-
-def _permission_snapshot(table, strategy_id: str) -> Dict[str, Any]:
-    resp = table.query(
-        KeyConditionExpression=Key("strategy_id").eq(strategy_id),
-        ProjectionExpression="principal",
-    )
-    principals = [item.get("principal") for item in resp.get("Items", [])]
-    public = any(principal == "ROLE#PUBLIC" for principal in principals)
-    fund = any(principal == "ROLE#FUND" for principal in principals)
-    netids = [
-        principal.replace("USER#", "", 1)
-        for principal in principals
-        if isinstance(principal, str) and principal.startswith("USER#")
-    ]
-
-    users: List[Dict[str, Any]] = []
-    for netid in netids:
-        resp = USERS_TABLE.get_item(Key={"netid": netid})
-        item = resp.get("Item") or {}
-        users.append(
-            {
-                "netid": netid,
-                "full_name": item.get("full_name") or "",
-                "uconn_email": item.get("uconn_email") or "",
-            }
-        )
-
-    return {"public": public, "fund": fund, "users": users}
-
-
-def _apply_permission_changes(
-    table,
-    strategy_id: str,
-    scope_body: Dict[str, Any],
-    mode: str,
-) -> Optional[Dict[str, Any]]:
-    public = scope_body.get("public")
-    fund = scope_body.get("fund")
-    add_users = scope_body.get("addUsers", [])
-    remove_users = scope_body.get("removeUsers", [])
-
-    if public is not None and not isinstance(public, bool):
-        return _json(400, {"message": "public must be a boolean"})
-    if fund is not None and not isinstance(fund, bool):
-        return _json(400, {"message": "fund must be a boolean"})
-    if add_users is not None and not isinstance(add_users, list):
-        return _json(400, {"message": "addUsers must be a list"})
-    if remove_users is not None and not isinstance(remove_users, list):
-        return _json(400, {"message": "removeUsers must be a list"})
-
-    if mode == "patch":
-        if public is True:
-            table.put_item(Item={"strategy_id": strategy_id, "principal": "ROLE#PUBLIC"})
-        elif public is False:
-            table.delete_item(Key={"strategy_id": strategy_id, "principal": "ROLE#PUBLIC"})
-        if fund is True:
-            table.put_item(Item={"strategy_id": strategy_id, "principal": "ROLE#FUND"})
-        elif fund is False:
-            table.delete_item(Key={"strategy_id": strategy_id, "principal": "ROLE#FUND"})
-        for netid in add_users or []:
-            if isinstance(netid, str) and netid.strip():
-                principal = _principal_for_user(netid.strip())
-                table.put_item(Item={"strategy_id": strategy_id, "principal": principal})
-    elif mode == "delete":
-        if public is True:
-            table.delete_item(Key={"strategy_id": strategy_id, "principal": "ROLE#PUBLIC"})
-        if fund is True:
-            table.delete_item(Key={"strategy_id": strategy_id, "principal": "ROLE#FUND"})
-        for netid in remove_users or []:
-            if isinstance(netid, str) and netid.strip():
-                principal = _principal_for_user(netid.strip())
-                table.delete_item(Key={"strategy_id": strategy_id, "principal": principal})
-    else:
-        return _json(500, {"message": "invalid mode"})
-
-    return None
-
-
-def _read_implied_changes(scope_body: Dict[str, Any]) -> Optional[Dict[str, Any]]:
-    implied: Dict[str, Any] = {}
-    if scope_body.get("public") is True:
-        implied["public"] = True
-    if scope_body.get("fund") is True:
-        implied["fund"] = True
-    add_users = scope_body.get("addUsers")
-    if isinstance(add_users, list) and add_users:
-        implied["addUsers"] = add_users
-    return implied or None
 
 
 # ----------------------------
@@ -579,50 +406,9 @@ def _get_netid_from_event(event: Dict[str, Any]) -> Optional[str]:
     return netid or None
 
 
-def _get_roles_from_event(event: Dict[str, Any]) -> List[str]:
-    request_context = event.get("requestContext") or {}
-    authorizer = request_context.get("authorizer") or {}
-
-    roles_raw = authorizer.get("roles")
-    if not roles_raw and isinstance(authorizer.get("lambda"), dict):
-        roles_raw = authorizer.get("lambda", {}).get("roles")
-
-    if isinstance(roles_raw, str):
-        try:
-            roles = json.loads(roles_raw)
-        except json.JSONDecodeError:
-            roles = []
-    elif isinstance(roles_raw, list):
-        roles = roles_raw
-    else:
-        roles = []
-
-    return [str(role) for role in roles if isinstance(role, (str, int))]
-
-
-def _has_read_permission(strategy_id: str, netid: str, roles: List[str]) -> bool:
-    principal = _principal_for_user(netid)
-    for table in (READ_PERMISSIONS_DDB, WRITE_PERMISSIONS_DDB):
-        resp = table.get_item(Key={"strategy_id": strategy_id, "principal": principal})
-        if "Item" in resp:
-            return True
-        public_resp = table.get_item(
-            Key={"strategy_id": strategy_id, "principal": "ROLE#PUBLIC"}
-        )
-        if "Item" in public_resp:
-            return True
-        for role_principal in _role_principals(roles):
-            role_resp = table.get_item(
-                Key={"strategy_id": strategy_id, "principal": role_principal}
-            )
-            if "Item" in role_resp:
-                return True
-    return False
-
-
 def _list_readable_strategy_ids(netid: str, roles: List[str]) -> List[str]:
     principals = [_principal_for_user(netid), "ROLE#PUBLIC"]
-    principals.extend(_role_principals(roles))
+    principals.extend(role_principals(roles))
     strategy_ids = _list_strategy_ids_for_principals(READ_PERMISSIONS_DDB, principals)
     write_ids = _list_strategy_ids_for_principals(WRITE_PERMISSIONS_DDB, principals)
     for sid in write_ids:
@@ -660,18 +446,6 @@ def _list_strategy_ids_for_principals(
                 break
 
     return strategy_ids
-
-
-def _role_principals(roles: List[str]) -> List[str]:
-    principals: List[str] = []
-    for role in roles:
-        if not isinstance(role, str):
-            continue
-        role = role.strip()
-        if not role or role == "ADMIN":
-            continue
-        principals.append(f"ROLE#{role}")
-    return principals
 
 
 def _chunk(items: List[Dict[str, Any]], size: int) -> List[List[Dict[str, Any]]]:
